moves the note and interval 'dbs' into a separate package
