+++ app/models/chords/ChordClasses.scala
@@ -7,7 +7,14 @@ package models.chords
 object ChordClasses extends Enumeration {
   type ChordClass = ChordClasses.Value
 
-  val minorTriad, majorTriad, diminishedTriad, augmentedTriad, minSixChord, majSixChord, minSixFour, majSixFour = Value
+  val minorTriad,
+  majorTriad,
+  diminishedTriad,
+  augmentedTriad,
+  minSixChord,
+  majSixChord,
+  minSixFour,
+  majSixFour = Value
 
 
 }
